Add more controls and work with domain account
I added more variables in the recipe.

I added attributes into the ERB file so we don't have to update the
variable block inside the template resource in the Chef recipe as we
more variables into the buildAgent.properties file.

Add a Powershell script to that lets SC.exe to add rights for a domain
user to a service.